+++ packages/swapkit/types/src/wallet.ts
@@ -135,15 +135,5 @@ export const availableChainsByWallet: any = {
 
 //prefurred chains for wallets (for large wallets this set help performance)
 export const prefurredChainsByWallet: any = {
-  [WalletOption.KEEPKEY]: [
-    Chain.Bitcoin,
-    Chain.BitcoinCash,
-    Chain.Ripple,
-    Chain.Dogecoin,
-    Chain.THORChain,
-    Chain.Ethereum,
-    Chain.Base,
-    Chain.Dash,
-    Chain.Arbitrum,
-  ],
+  [WalletOption.KEEPKEY]: [Chain.Bitcoin, Chain.Ethereum, Chain.Base],
 };