+++ packages/swapkit/entities/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @thorswap-lib/swapkit-entities
 
+## 3.7.28
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.28
+
+## 3.7.27
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.27
+
+## 3.7.26
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.26
+
+## 3.7.25
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.25
+
 ## 3.7.24
 
 ### Patch Changes