+++ packages/swapkit/helpers/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @swapkit/helpers
 
+## 11.0.28
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/tokens@3.7.28
+  - @coinmasters/types@4.7.28
+
+## 11.0.27
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/tokens@3.7.27
+  - @coinmasters/types@4.7.27
+
+## 11.0.26
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/tokens@3.7.26
+  - @coinmasters/types@4.7.26
+
+## 11.0.25
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/tokens@3.7.25
+  - @coinmasters/types@4.7.25
+
 ## 11.0.24
 
 ### Patch Changes