+++ packages/swapkit/types/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @swapkit/types
 
+## 4.7.28
+
+### Patch Changes
+
+- bump
+
+## 4.7.27
+
+### Patch Changes
+
+- bump
+
+## 4.7.26
+
+### Patch Changes
+
+- bump
+
+## 4.7.25
+
+### Patch Changes
+
+- bump
+
 ## 4.7.24
 
 ### Patch Changes